+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Patch Drupal core to fix [Issue #3266341: Views pagers do math on disparate data types, resulting in type errors in PHP 8](https://www.drupal.org/project/drupal/issues/3266341)
 - [Implement FarmBreadcrumbBuilder::applies() to only affect desired routes](https://github.com/farmOS/farmOS/pull/644)
 - [Attempt to update the map size until it is rendered #576](https://github.com/farmOS/farmOS/issues/576)
+- [Show map popup for the smallest feature at the clicked point #652](https://github.com/farmOS/farmOS/pull/652)
 
 ## [2.0.1] 2023-02-08